refactor: break down large app_test.go file into focused test files
Refactored internal/thinktank/app_test.go (1158 lines) into smaller,
more maintainable files to comply with 1000-line limit:

- app_test_mocks.go: Mock implementations (MockAuditLogger, MockLogger, etc.)
- app_test_helpers.go: Test utility functions (setupTestEnvironment, createTestFile)
- app_execute_test.go: Happy path and dry run tests
- app_error_test.go: Error handling test scenarios
- app_test.go: Minimal file with documentation of refactoring

All tests continue to pass and 90.4% coverage is maintained.
